app/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
coverage==5.5
deprecation==2.1.0
Django==3.2.8
django-cors-headers==3.10.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
eyed3==0.9.6
filetype==1.0.8
idna==3.2
packaging==21.0
pyparsing==2.4.7
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
